move schema.js to the `contexts` dir
