+++ src/attribute.js
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import { LocationContext } from './contexts/location';
 import { hasSetEntry } from './utils';
-import { SchemaContext } from './schema';
+import { SchemaContext } from './contexts/schema';
 
 const Attribute = ({ attribute, type, includeEnabled }) => {
   const { forPath } = useContext(SchemaContext);
